# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PUFFIN SRL - Plataforma Integral de Gestión Operativa
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /maquinas:
    get:
      operationId: getMaquinas
      tags: [maquinas]
      summary: Listar maquinaria
      parameters:
        - name: estado
          in: query
          schema:
            type: string
            enum: [activa, detenida, mantenimiento, baja]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Lista de maquinaria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Maquina"
    post:
      operationId: createMaquina
      tags: [maquinas]
      summary: Registrar maquinaria
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaquinaInput"
      responses:
        "201":
          description: Maquinaria creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Maquina"
  /maquinas/{id}:
    get:
      operationId: getMaquina
      tags: [maquinas]
      summary: Obtener maquinaria por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Maquinaria
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Maquina"
    put:
      operationId: updateMaquina
      tags: [maquinas]
      summary: Actualizar maquinaria
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaquinaUpdate"
      responses:
        "200":
          description: Maquinaria actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Maquina"
  /jornadas:
    get:
      operationId: getJornadas
      tags: [jornadas]
      summary: Listar jornadas
      parameters:
        - name: empleado_id
          in: query
          schema:
            type: integer
        - name: maquina_id
          in: query
          schema:
            type: integer
        - name: fecha_desde
          in: query
          schema:
            type: string
        - name: fecha_hasta
          in: query
          schema:
            type: string
        - name: estado
          in: query
          schema:
            type: string
            enum: [en_curso, finalizada]
      responses:
        "200":
          description: Lista de jornadas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Jornada"
  /jornadas/iniciar:
    post:
      operationId: iniciarJornada
      tags: [jornadas]
      summary: Iniciar jornada de trabajo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IniciarJornadaInput"
      responses:
        "201":
          description: Jornada iniciada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Jornada"
  /jornadas/{id}/finalizar:
    post:
      operationId: finalizarJornada
      tags: [jornadas]
      summary: Finalizar jornada de trabajo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinalizarJornadaInput"
      responses:
        "200":
          description: Jornada finalizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Jornada"
...
  /mantenimientos:
    get:
      operationId: getMantenimientos
      tags: [mantenimientos]
      summary: Listar mantenimientos
      parameters:
        - name: maquina_id
          in: query
          schema:
            type: integer
        - name: tipo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Lista de mantenimientos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Mantenimiento"
    post:
      operationId: createMantenimiento
      tags: [mantenimientos]
      summary: Registrar mantenimiento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MantenimientoInput"
      responses:
        "201":
          description: Mantenimiento registrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mantenimiento"
...
components:
  schemas:
...
    Maquina:
      type: object
      required: [id, nombre, tipo, estado]
      properties:
        id:
          type: integer
        empresa_id:
          type: integer
        codigo:
          type: ["string", "null"]
        nombre:
          type: string
        tipo:
          type: string
        marca:
          type: ["string", "null"]
        modelo:
          type: ["string", "null"]
        anio:
          type: ["integer", "null"]
        patente:
          type: ["string", "null"]
        dominio:
          type: ["string", "null"]
        chasis:
          type: ["string", "null"]
        motor:
          type: ["string", "null"]
        horometro:
          type: number
        kilometros:
          type: number
        estado:
          type: string
          enum: [activa, detenida, mantenimiento, baja]
        ultimo_service:
          type: ["string", "null"]
        proximo_service:
          type: ["string", "null"]
    MaquinaInput:
      type: object
      required: [nombre, tipo]
      properties:
        codigo:
          type: string
        nombre:
          type: string
        tipo:
          type: string
        marca:
          type: string
        modelo:
          type: string
        anio:
          type: integer
        patente:
          type: string
        dominio:
          type: string
        chasis:
          type: string
        motor:
          type: string
        horometro:
          type: number
        kilometros:
          type: number
    MaquinaUpdate:
      type: object
      properties:
        nombre:
          type: string
        estado:
          type: string
          enum: [activa, detenida, mantenimiento, baja]
        horometro:
          type: number
        kilometros:
          type: number
        proximo_service:
          type: string
    Jornada:
      type: object
      required: [id, empleado_id, maquina_id, fecha, estado]
      properties:
        id:
          type: integer
        empresa_id:
          type: integer
        empleado_id:
          type: integer
        maquina_id:
          type: integer
        empleado_nombre:
          type: string
        maquina_nombre:
          type: string
        fecha:
          type: string
        hora_inicio:
          type: ["string", "null"]
        hora_fin:
          type: ["string", "null"]
        km_inicio:
          type: ["number", "null"]
        km_fin:
          type: ["number", "null"]
        horometro_inicio:
          type: ["number", "null"]
        horometro_fin:
          type: ["number", "null"]
        observaciones:
          type: ["string", "null"]
        problemas:
          type: ["string", "null"]
        estado:
          type: string
          enum: [en_curso, finalizada]
        horas_trabajadas:
          type: ["number", "null"]
    IniciarJornadaInput:
      type: object
      required: [empleado_id, maquina_id, horometro_inicio]
      properties:
        empleado_id:
          type: integer
        maquina_id:
          type: integer
        horometro_inicio:
          type: number
        km_inicio:
          type: number
        observaciones:
          type: string
    FinalizarJornadaInput:
      type: object
      required: [horometro_fin]
      properties:
        horometro_fin:
          type: number
        km_fin:
          type: number
        problemas:
          type: string
...
    Mantenimiento:
      type: object
      required: [id, maquina_id, fecha, tipo]
      properties:
        id:
          type: integer
        empresa_id:
          type: integer
        maquina_id:
          type: integer
        maquina_nombre:
          type: string
        fecha:
          type: string
        horas:
          type: ["number", "null"]
        tipo:
          type: string
        descripcion:
          type: ["string", "null"]
        proximo_service:
          type: ["string", "null"]
        estado:
          type: string
          enum: [pendiente, realizado, cancelado]
    MantenimientoInput:
      type: object
      required: [maquina_id, tipo]
      properties:
        maquina_id:
          type: integer
        horas:
          type: number
        tipo:
          type: string
        descripcion:
          type: string
        proximo_service:
          type: string
...