# lib/api-zod/src/generated/types/finalizarJornadaInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * PUFFIN SRL - Plataforma Integral de Gestión Operativa
 * OpenAPI spec version: 0.1.0
 */

export interface FinalizarJornadaInput {
  horometro_fin: number;
  km_fin?: number;
  problemas?: string;
}