# lib/api-zod/src/generated/types/mantenimiento.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * PUFFIN SRL - Plataforma Integral de Gestión Operativa
 * OpenAPI spec version: 0.1.0
 */
import type { MantenimientoEstado } from './mantenimientoEstado';

export interface Mantenimiento {
  id: number;
  empresa_id?: number;
  maquina_id: number;
  maquina_nombre?: string;
  fecha: string;
  /** @nullable */
  horas?: number | null;
  tipo: string;
  /** @nullable */
  descripcion?: string | null;
  /** @nullable */
  proximo_service?: string | null;
  estado?: MantenimientoEstado;
}